+++ src/components/QuickTextBox/Messages.js
@@ -1,0 +1,11 @@
+import { defineMessages } from 'react-intl'
+
+/**
+ * Internationalized messages for use with QuickTextBox
+ */
+export default defineMessages({
+  saveLabel: {
+    id: "QuickTextBox.controls.save.label",
+    defaultMessage: "Save",
+  },
+})