+++ src/components/SavedTasks/SavedTasksWidget.js
@@ -1,12 +1,20 @@
 import React, { Component } from 'react'
 import { FormattedMessage } from 'react-intl'
+import classNames from 'classnames'
 import _map from 'lodash/map'
 import _get from 'lodash/get'
 import _compact from 'lodash/compact'
 import _isFinite from 'lodash/isFinite'
+import _kebabCase from 'lodash/kebabCase'
 import { Link } from 'react-router-dom'
 import { WidgetDataTarget, registerWidgetType }
        from '../../services/Widget/Widget'
+import { keysByStatus, messagesByStatus }
+       from '../../services/Task/TaskStatus/TaskStatus'
+import { keysByReviewStatus, messagesByReviewStatus }
+       from '../../services/Task/TaskReview/TaskReviewStatus'
+import TaskCommentsModal
+       from '../../components/TaskCommentsModal/TaskCommentsModal'
 import QuickWidget from '../QuickWidget/QuickWidget'
 import Dropdown from '../Dropdown/Dropdown'
 import SvgSymbol from '../SvgSymbol/SvgSymbol'
@@ -25,6 +33,10 @@ const descriptor = {
 }
 
 export default class SavedTasksWidget extends Component {
+  state = {
+    openComments: null,
+  }
+
   componentDidMount() {
     if (this.props.user && this.props.fetchSavedTasks) {
       this.props.fetchSavedTasks(this.props.user.id)
@@ -38,7 +50,16 @@ export default class SavedTasksWidget extends Component {
         className="saved-tasks-widget"
         widgetTitle={<FormattedMessage {...messages.header} />}
       >
-        <SavedTaskList {...this.props} />
+        <SavedTaskList
+          {...this.props}
+          openComments={taskId => this.setState({openComments: taskId})}
+        />
+        {_isFinite(this.state.openComments) &&
+         <TaskCommentsModal
+           taskId={this.state.openComments}
+           onClose={() => this.setState({openComments: null})}
+         />
+        }
       </QuickWidget>
     )
   }
@@ -52,14 +73,29 @@ const SavedTaskList = function(props) {
       }
 
       return (
-        <li
-          key={task.id}
-          className="mr-h-5 mr-my-2 mr-flex mr-justify-between mr-items-center"
-        >
-          <Link to={`/challenge/${task.parent.id}/task/${task.id}`}>
-            <span className="saved-tasks__task__name">{task.name}</span> &mdash; {task.parent.name}
-          </Link>
-          <div className="mr-h-5">
+        <React.Fragment key={task.id}>
+          <div className="mr-col-span-5">
+            <Link to={`/challenge/${task.parent.id}/task/${task.id}`}>{task.name}</Link>
+            <div className="mr-links-grey-light">
+              <Link to={`/browse/challenges/${task.parent.id}`}>{task.parent.name}</Link>
+            </div>
+          </div>
+          <div className={classNames(
+            "mr-col-span-3",
+            `mr-status-${_kebabCase(keysByStatus[task.status])}`
+          )}>
+            <FormattedMessage {...messagesByStatus[task.status]} />
+          </div>
+          <div className={classNames(
+            "mr-col-span-3",
+            `mr-review-${_kebabCase(keysByReviewStatus[task.reviewStatus])}`
+          )}>
+            {_isFinite(task.reviewStatus) ?
+            <FormattedMessage {...messagesByReviewStatus[task.reviewStatus]} /> :
+            <span />
+            }
+          </div>
+          <div className="mr-h-5 mr-text-right">
             <Dropdown
               className="mr-dropdown--right"
               dropdownButton={dropdown => (
@@ -83,6 +119,12 @@ const SavedTaskList = function(props) {
                   </li>
                   <li>
                     {/* eslint-disable-next-line jsx-a11y/anchor-is-valid */}
+                    <a onClick={() => props.openComments(task.id)}>
+                      <FormattedMessage {...messages.viewComments} />
+                    </a>
+                  </li>
+                  <li>
+                    {/* eslint-disable-next-line jsx-a11y/anchor-is-valid */}
                     <a onClick={() => props.unsaveTask(props.user.id, task.id)}>
                       <FormattedMessage {...messages.unsave} />
                     </a>
@@ -91,16 +133,16 @@ const SavedTaskList = function(props) {
               }
             />
           </div>
-        </li>
+        </React.Fragment>
       )
     }
   ))
 
   return (
     taskItems.length > 0 ?
-    <ol className="mr-list-reset mr-links-green-lighter mr-pb-24">
+    <div className="mr-links-green-lighter mr-grid mr-grid-columns-12 mr-grid-gap-4 mr-pb-24 mr-justify-between mr-pt-1">
       {taskItems}
-    </ol> :
+    </div> :
     <div className="mr-text-grey-lighter">
       <FormattedMessage {...messages.noTasks} />
     </div>